% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{circshift}
\alias{circshift}
\title{Circular shift of array x by n steps.}
\usage{
circshift(x, n)
}
\arguments{
\item{x}{An array.}

\item{n}{Number of steps for circular shift (positive for forward shift, negative for backward shift)}
}
\value{
Circular-shifted x.
}
\description{
Circular shift of array x by n steps.
}

