% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{discretize_letter}
\alias{discretize_letter}
\title{Discretize matrix according to letter.}
\usage{
discretize_letter(x, letter)
}
\arguments{
\item{x}{A data matrix.}

\item{letter}{A letter for discretizing.}
}
\value{
A binary matrix.
}
\description{
Given a discrete matrix, discretize it so that any element matching
letter is set to one and otherwise to zero.
}

