% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_global_tile.R
\name{dp_segment}
\alias{dp_segment}
\title{Segmentation using dynamic programming}
\usage{
dp_segment(get_tile, m, K)
}
\arguments{
\item{get_tile}{Function used to get tile.}

\item{m}{The width of the matrix A, or more generally, the rightmost index where the segmentation can end.}

\item{K}{is the number of segments.}
}
\value{
Something
}
\description{
Segmentation using dynamic programming
}

