% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_global_tile.R
\name{find_global_tiles}
\alias{find_global_tiles}
\title{Find global tiles}
\usage{
find_global_tiles(X, num_tiles, precision_lowerbound)
}
\arguments{
\item{X}{the data matrix}

\item{num_tiles}{the number of tiles}

\item{precision_lowerbound}{the lower bound on precision}
}
\value{
The tiling
}
\description{
Find global tiles
}

