% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_global_tile.R
\name{get_segment_bounds}
\alias{get_segment_bounds}
\title{Get segment bounds}
\usage{
get_segment_bounds(TI, K)
}
\arguments{
\item{TI}{Something}

\item{K}{is the number of segments.}
}
\value{
The start positions of the segments in an optimal segmentation to K segments.
}
\description{
Get segment bounds
}

