% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_global_tile.R
\name{global_tile}
\alias{global_tile}
\title{Find tiles using the GlobalTile algorithm.}
\usage{
global_tile(X, K = 1, alpha = 0.75)
}
\arguments{
\item{X}{The data matrix.}

\item{K}{The maximum number of tiles. Default is 1.}

\item{alpha}{The precision threshold. Default is 0.75.}
}
\value{
All tiles with a precision of at least alpha.
}
\description{
This function finds tiles using the GlobalTile algorithm.
}

