% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_01model}
\alias{make_01model}
\title{Generates a random 0-1 matrix with some randomly planted
segments. Every planted segment has a single subset of rows filled
with 1s with probability p. The subset has floor(n/l)
rows. Otherwise the matrix is filled with 0s.}
\usage{
make_01model(n = 50, m = 100, k = 5, p = 0.8, l = 5)
}
\arguments{
\item{n}{Number of rows.}

\item{m}{Number of columns.}

\item{k}{Number of randomly planted segments.}

\item{p}{Probability of 1s in the segments.}

\item{l}{Span of segment.}
}
\value{
A binary data matrix with k segments.
}
\description{
It might be better in practice to set m to a large enough value,
like O(100) or so. The default values give a nice example.
}

