% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_01tiles}
\alias{make_01tiles}
\title{Generates a random 0-1 matrix with n rows and m columns, and k
randomly placed "tiles". Every planted tile has a single subset of
l rows that span l consecutive columns.  The matrix is filled with
zeros, 1s occur in tiles with probability p.}
\usage{
make_01tiles(n = 50, m = 100, k = 5, p = 0.8, l = 20)
}
\arguments{
\item{n}{Number of rows.}

\item{m}{Number of columns.}

\item{k}{Number of randomly planted tiles.}

\item{p}{Probability of 1s in the segments.}

\item{l}{Span of tiles.}
}
\value{
A binary data matrix with k segments.
}
\description{
Generates a random 0-1 matrix with n rows and m columns, and k
randomly placed "tiles". Every planted tile has a single subset of
l rows that span l consecutive columns.  The matrix is filled with
zeros, 1s occur in tiles with probability p.
}

