% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_dataset_list_synthetic}
\alias{make_dataset_list_synthetic}
\title{Make list with synthetic datasets}
\usage{
make_dataset_list_synthetic()
}
\value{
List with synthetic datasets
}
\description{
Make list with synthetic datasets
}

