% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_tile}
\alias{make_tile}
\title{Make a tile}
\usage{
make_tile(rows, colstart, colstop, tile_id = 0)
}
\arguments{
\item{rows}{The rows in the tile}

\item{colstart}{The starting column of the tile}

\item{colstop}{The ending column o fthe tile}

\item{tile_id}{The ID of the tile. Defaults to zero.}
}
\value{
A one-row matrix with the information needed
}
\description{
Make a tile
}

