% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_max_tile.R
\name{max_tile_main}
\alias{max_tile_main}
\title{Find maximum tiles given precision threshold alpha.}
\usage{
max_tile_main(X, K = 1, alpha = 0.75, rowwise = FALSE, mem = FALSE)
}
\arguments{
\item{alpha}{The precision threshold}

\item{rowwise}{Proceed rowwise (boolean).}

\item{mem}{Memoise (boolean).}

\item{x}{The data matrix.}
}
\value{
All tiles with a precision of at least alpha.
}
\description{
This function finds all maximum tiles in the data matrix x having a precision of at least alpha.
}

