% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_global_tile.R
\name{max_weight_tile_fnc}
\alias{max_weight_tile_fnc}
\title{Max weight tile function}
\usage{
max_weight_tile_fnc(data_matrix, prec_lowerbound)
}
\arguments{
\item{data_matrix}{the data matrix}

\item{prec_lwoerbound}{lower bound on the precision}
}
\value{
Something
}
\description{
Returns a function that takes as arguments the left and right boundary of a tile,
and returns a list with components
R: the rows that belong to the tile
W: the weight of the tile.
The function should be used to extract the maximum weight tile
from data_matrix with given left and right boundaries
subject to prec_lowerbound.
}

