% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perf_all_metrics}
\alias{perf_all_metrics}
\title{Calculate all performance metrics.}
\usage{
perf_all_metrics(x, tiles, print = FALSE, as_string = FALSE,
  short_names = TRUE)
}
\arguments{
\item{x}{A data matrix.}

\item{tiles}{A tiling.}
}
\value{
A list with all of the performance metrics
}
\description{
Calculate all performance metrics.
}

