% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perf_cardinality}
\alias{perf_cardinality}
\title{Calculate cardinality.}
\usage{
perf_cardinality(tiles, x)
}
\arguments{
\item{tiles}{A tiling.}

\item{x}{A data matrix (not used).}
}
\value{
The cardinality (integer).
}
\description{
Calculate the cardinality as the number of unique tiles (clusters) in the tiling.
}

