% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perf_precision}
\alias{perf_precision}
\title{Calculate precision.}
\usage{
perf_precision(tiles, x)
}
\arguments{
\item{tiles}{A tiling.}

\item{x}{A data matrix.}
}
\value{
The precision.
}
\description{
Calculate precision (fraction of ones contained in the tiles).
}

