% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perf_recall}
\alias{perf_recall}
\title{Calculate recall.}
\usage{
perf_recall(tiles, x, xsum = sum(x))
}
\arguments{
\item{tiles}{A tiling.}

\item{x}{A data matrix.}
}
\value{
The recall.
}
\description{
Calculate recall (the fraction of ones in a binary matrix covered by a tiling).
}

