% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{plot_binary_matrix}
\alias{plot_binary_matrix}
\title{Plot a binary matrix optionally with tiles.}
\usage{
plot_binary_matrix(x, tiles = NULL, tiles_scaled = NULL, xsignal = NULL,
  plot_data = TRUE, plot_discrete = TRUE, plot_signal = FALSE,
  plot_dot = FALSE, main = "", savename = NULL, plot_y_labels = TRUE,
  xlab = "", ylab = "", perf_string = TRUE, skipxlabel = 30,
  plot_x_labels = FALSE, xvec = NULL, pointcol = "black", cex.ylab = 1,
  cex.main = 2, cex.axis = 2, cex.xlab = 1)
}
\arguments{
\item{x}{Data matrix.}

\item{tiles}{A matrix with tiles.}

\item{plot_data}{Should the data be plotted (Boolean, default is TRUE).}

\item{plot_dot}{Should a dot be plotted for a zero (Boolean, default is FALSE).}

\item{main}{Plot title.}
}
\value{
Nothing
}
\description{
Plot a binary matrix optionally with tiles. A bit is plotted as an
asterisk and a zero as a dot (or optionally not plotted).
}

