% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_global_tile.R
\name{plot_segments}
\alias{plot_segments}
\title{Plot segments}
\usage{
plot_segments(A, bounds, ...)
}
\arguments{
\item{A}{a matrix generated e.g. by the make_01model function.}

\item{bounds}{the list of bounds as returned by the dp_segment function.}
}
\value{
Nothing
}
\description{
Plot the input matrix together with the segment boundaries.
}
\details{
Takes the bounds output of the dynamic programming algboritm
and converts this to a list of tiles. Every tile is a list
with components R, a, b, and W, where W is the weight of the tile.
}

