% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{random_shift_matrix}
\alias{random_shift_matrix}
\title{Shift n randomly chosen rows from the matrix by steps.}
\usage{
random_shift_matrix(x, n, steps = 0, steps_low = NULL, steps_high = NULL,
  random = FALSE)
}
\arguments{
\item{x}{An array.}

\item{n}{Number of steps for circular shift.}
}
\value{
Circular-shifted x.
}
\description{
Shift n randomly chosen rows from the matrix by steps.
}

