% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{segmat_to_list}
\alias{segmat_to_list}
\title{Convert a matrix with segments to a list}
\usage{
segmat_to_list(segmat)
}
\arguments{
\item{segmat}{A matrix with segments}
}
\value{
A list containing the tiles
}
\description{
Convert a matrix with segments to a list
}

