% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_global_tile.R
\name{segments_to_tiles}
\alias{segments_to_tiles}
\title{Segments to tiles}
\usage{
segments_to_tiles(X, bounds, get_tile)
}
\arguments{
\item{X}{the data matrix}

\item{bounds}{contains the start positions of the segments.}

\item{get_tile}{Something}
}
\value{
Tiles
}
\description{
Takes the bounds output of the dynamic programming algboritm
and converts this to a list of tiles. Every tile is a list
with components R, a, b, and W, where W is the weight of the tile.
}

