% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{shuffle_matrix}
\alias{shuffle_matrix}
\title{Randomly shuffle each row of a matrix}
\usage{
shuffle_matrix(x)
}
\arguments{
\item{x}{An array.}
}
\value{
A shuffled vesion of x.
}
\description{
Randomly shuffle each row of a matrix
}

