% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sub2ind}
\alias{sub2ind}
\title{Subscript to linear index}
\usage{
sub2ind(row, col, nrows)
}
\arguments{
\item{row}{Row index}

\item{col}{Column indexd}

\item{nrows}{The number of rows in the data matrix}
}
\value{
The linear index corresponding to the subscript index (row, col)
}
\description{
Subscript to linear index
}

