% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tile2ind}
\alias{tile2ind}
\title{Convert a tile to linear matrix indices}
\usage{
tile2ind(tile, m_dim)
}
\arguments{
\item{tile}{The tile}

\item{m_dim}{The dimensions of the data matrix}
}
\value{
Linear indices corresponding to the tile
}
\description{
Convert a tile to linear matrix indices
}

