% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tiles2ind}
\alias{tiles2ind}
\title{Convert list of tiles to linear matrix indices}
\usage{
tiles2ind(tiles, m_dim)
}
\arguments{
\item{tiles}{List of tiles}

\item{m_dim}{The dimensions of the data matrix}
}
\value{
List with linear indices corresponding to the tiles
}
\description{
Convert list of tiles to linear matrix indices
}

